requirements.txt
numpy==1.23.4
Pillow==9.3.0
Pillow==10.1.0
PyAutoGUI==0.9.54
